requirements.txt
django
requests
kafka-python